# Phase 5: WebGPU Mock Optimization and Performance Monitor Test Stabilization