## Summary
- Enhanced WebGPU mock system with comprehensive device and adapter simulation
- Fixed all Performance Monitor tests (137 tests now passing)
- Improved test success rate from 76.3% to 77.0% (+9 tests)
- Resolved navigator undefined errors and test timing issues

## Key Improvements

### WebGPU Mock System
- **Enhanced mock adapter** with realistic limits and features simulation
- **Comprehensive device mocking** including buffer, texture, and pipeline operations
- **Proper command encoder simulation** with render pass and compute pass support
- **Scenario-based testing** (no-adapter, slow-adapter scenarios)
- **Memory usage simulation** for performance monitoring integration

### Performance Monitor Fixes
- **Fixed navigator undefined errors** in AdaptiveQualityManager with proper checks
- **Resolved core test issues** with history access (entries.length vs length)
- **Enhanced scenarios tests** with proper async timing and realistic mocking
- **Improved integration test stability** with better mock isolation
- **Fixed AudioCollector tests** with proper createOscillator and addEventListener mocking

### Test Infrastructure Improvements
- **Added comprehensive test utilities** in `src/utils/test/`
  - `webgpuMocks.ts`: Complete WebGPU device and adapter simulation
  - `performanceMonitorMocks.ts`: Performance metrics and collector mocks
  - `testSetup.ts`: Enhanced test environment setup
- **Enhanced jest.setup.js** with better WebGPU and performance mocks
- **Improved dynamic import handling** and error simulation scenarios
- **Better mock cleanup** and isolation between test runs

## Test Results Impact

| Metric | Before | After | Change |
|--------|--------|--------|--------|
| **Tests Passing** | 944 | 953 | +9 tests ✅ |
| **Tests Failing** | 119 | 110 | -9 tests ✅ |
| **Success Rate** | 76.3% | 77.0% | +0.7% ✅ |
| **Performance Monitor Tests** | Mixed | 137/137 | 🎯 100% |

## Technical Details

### WebGPU Mock Architecture
- **Realistic adapter simulation** with proper feature sets and limits
- **Device lifecycle management** with proper initialization and cleanup
- **Buffer and texture operations** with size and usage validation
- **Pipeline creation** for both render and compute operations
- **Command submission** with proper state tracking

### Performance Monitor Integration
- **Navigator compatibility** checks for cross-environment testing
- **Memory monitoring** with dynamic heap size simulation
- **Device capability detection** with proper fallback mechanisms
- **Quality profile management** with automatic adaptation logic

### Code Quality Improvements
- **Enhanced error handling** in dynamic import scenarios
- **Better mock isolation** preventing cross-test contamination
- **Improved async timing** in scenario and integration tests
- **Type safety improvements** while maintaining mock flexibility

## Test Categories Fixed
- ✅ **AdaptiveQualityManager** (15/15 tests passing)
- ✅ **Performance Monitor Core** (28/28 tests passing) 
- ✅ **Performance Monitor Collectors** (44/44 tests passing)
- ✅ **Performance Monitor Scenarios** (8/8 tests passing)
- ✅ **Performance Monitor Integration** (15/15 tests passing)
- ✅ **Performance Monitor Types** (27/27 tests passing)

## Next Phase Preparation
This establishes a solid foundation for Phase 6, which will target:
- Additional mock system improvements for remaining failing tests
- Enhanced error scenario coverage
- Further test stability improvements toward the 95% success rate goal

🤖 Generated with [Claude Code](https://claude.ai/code)